+++ src/app/storage/storage.component.ts
@@ -1,12 +1,36 @@
-import angular from 'angular';
-import _ from 'underscore';
+import { StateParams, StateService } from '@uirouter/angularjs';
+import {
+  element,
+  IQService,
+  IScope,
+  ITimeoutService,
+  IWindowService
+  } from 'angular';
+import * as _ from 'underscore';
+import { reportException } from '../exceptions';
 import { sum } from '../util';
-
 import template from './storage.html';
 import './storage.scss';
-import { reportException } from '../exceptions';
+import { SyncServiceType, StorageAdapter } from './sync.service';
 
-function StorageController($scope, dimSettingsService, SyncService, GoogleDriveStorage, dimDestinyTrackerService, $timeout, $window, $q, $i18next, $stateParams, $state) {
+declare global {
+  interface Window {
+    MSStream: any;
+  }
+}
+
+function StorageController(
+  $scope: IScope,
+  dimSettingsService,
+  SyncService: SyncServiceType,
+  dimDestinyTrackerService,
+  $timeout: ITimeoutService,
+  $window: IWindowService,
+  $q: IQService,
+  $i18next,
+  $stateParams: StateParams,
+  $state: StateService
+) {
   'ngInject';
 
   const vm = this;
@@ -23,7 +47,7 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
   function dataStats(data) {
     const taggedItems = sum(Object.keys(data)
                             .filter((k) => k.startsWith('dimItemInfo'))
-                            .map((k) => _.size(data[k])));
+                            .map((k) => _.size(data[k])), (v) => v);
 
     return {
       Loadouts: _.size(data['loadouts-v3.0']),
@@ -33,17 +57,13 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
     };
   }
 
-  function refreshAdapter(adapter) {
+  function refreshAdapter(adapter: StorageAdapter) {
     if (adapter.enabled) {
-      return adapter.get()
+      return $q.when(adapter.get())
         .then((data) => {
-          if (data) {
-            vm.adapterStats[adapter.name] = dataStats(data);
-          } else {
-            vm.adapterStats[adapter.name] = null;
-          }
+          vm.adapterStats[adapter.name] = data ? dataStats(data) : null;
         })
-        .catch((e) => {
+        .catch(() => {
           vm.adapterStats[adapter.name] = null;
         });
     } else {
@@ -58,15 +78,15 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
     });
   });
 
-  vm.forceSync = function() {
+  vm.forceSync = () => {
     return SyncService.get(true)
       .then((data) => SyncService.set(data, true))
       .then(() => $q.all(SyncService.adapters.map(refreshAdapter)));
   };
 
-  vm.driveSync = function() {
+  vm.driveSync = () => {
     if ($window.confirm($i18next.t('Storage.GDriveSignInWarning'))) {
-      return GoogleDriveStorage.authorize()
+      return SyncService.GoogleDriveStorage.authorize()
         .then(vm.forceSync)
         .catch((e) => {
           $window.alert($i18next.t('Storage.GDriveSignInError') + e.message);
@@ -76,16 +96,16 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
     return null;
   };
 
-  vm.driveLogout = function() {
+  vm.driveLogout = () => {
     $window.alert($i18next.t('Storage.GDriveLogout'));
-    return GoogleDriveStorage.revokeDrive();
+    return SyncService.GoogleDriveStorage.revokeDrive();
   };
 
-  vm.exportData = function() {
+  vm.exportData = () => {
     // Function to download data to a file
     function download(data, filename, type) {
       const a = document.createElement("a");
-      const file = new Blob([data], { type: type });
+      const file = new Blob([data], { type });
       const url = URL.createObjectURL(file);
       a.href = url;
       a.download = filename;
@@ -102,17 +122,17 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
     });
   };
 
-  vm.importData = function() {
+  vm.importData = () => {
     const reader = new FileReader();
-    reader.onload = function() {
+    reader.onload = () => {
       $scope.$apply(() => {
         // TODO: we're kinda trusting that this is the right data here, no validation!
         SyncService.set(JSON.parse(reader.result), true)
-          .then(() => $q.all(SyncService.adapters.forEach(refreshAdapter)));
+          .then(() => $q.all(SyncService.adapters.map(refreshAdapter)));
       });
       $window.alert($i18next.t('Storage.ImportSuccess'));
     };
-    const file = angular.element(document.getElementById('importFile'))[0].files[0];
+    const file = (element(document.getElementById('importFile')!)[0] as HTMLInputElement).files![0];
     if (file) {
       reader.readAsText(file);
     } else {
@@ -122,7 +142,7 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
 
   $scope.$on('gdrive-sign-in', () => {
     if ($stateParams.gdrive === 'true') {
-      vm.forceSync().then(() => $state.go('settings', { gdrive: undefined }, 'replace'));
+      vm.forceSync().then(() => $state.go('settings', { gdrive: undefined }, { location: 'replace' }));
     }
   });
 
@@ -133,7 +153,7 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
     });
   }
 
-  vm.clearIgnoredUsers = function() {
+  vm.clearIgnoredUsers = () => {
     if (!vm.canClearIgnoredUsers) {
       return;
     }
@@ -144,5 +164,5 @@ function StorageController($scope, dimSettingsService, SyncService, GoogleDriveS
 
 export const StorageComponent = {
   controller: StorageController,
-  template: template
+  template
 };
